+++ apps/docs/src/lib/getGithubEditUrl.ts
@@ -5,6 +5,6 @@ export const getGithubEditUrl = (Astro: Readonly<AstroGlobal>) => {
   const currentPage = Astro.url.pathname
   return `https://github.com/threlte/threlte/blob/main/apps/docs/src/content/${currentPage.replace(
     '/docs/',
-    '/'
+    ''
   )}.mdx`
 }