+++ packages/extras/src/lib/hooks/useTexture.ts
@@ -1,5 +1,6 @@
 import {
   useLoader,
+  useThrelte,
   type UseLoaderLoadInput,
   type UseLoaderLoadResult,
   type UseLoaderOptions
@@ -8,8 +9,25 @@ import { TextureLoader } from 'three'
 
 export const useTexture = <Input extends UseLoaderLoadInput>(
   input: Input,
-  options?: UseLoaderOptions<TextureLoader>
+  options?: UseLoaderOptions<TextureLoader> &
+    Parameters<ReturnType<typeof useLoader<typeof TextureLoader>>['load']>[1]
 ): UseLoaderLoadResult<TextureLoader, Input> => {
   const loader = useLoader(TextureLoader, options)
-  return loader.load(input)
+
+  const { renderer } = useThrelte()
+
+  return loader.load(input, {
+    ...options,
+    transform: (res) => {
+      if ('colorSpace' in res) {
+        // >= r152
+        res.colorSpace = renderer.outputColorSpace
+      } else {
+        // < r152
+        ;(res as any).encoding = (renderer as any).outputEncoding
+      }
+      res.needsUpdate = true
+      return options?.transform?.(res) ?? res
+    }
+  })
 }